IA-984 refactor LQAS/IM
- move CaregiversTable to components
- create LqasImMapHeader, delete LQAS- ImMapHeader components
- remove unused returned value from getLqas- getImStatsForRound utils
